+++ packages/sdk/src/modules/buildModule.ts
@@ -1,54 +1,22 @@
+/* eslint-disable no-redeclare */
 import mergeDeep from "../helpers/mergeDeep";
 import {
   Module,
   Extension,
-  ModuleInitializer,
   ExtensionInitializer,
   ModuleOptions,
+  ModuleInitializer,
+  ModuleInitializerWithMandatoryOptions,
 } from "../types";
 
-/* eslint-disable no-redeclare */
-function buildModule<
-  InitializedModule extends Module,
-  Options extends ModuleOptions = object
->(
-  module: ModuleInitializer<InitializedModule, Options>,
-  moduleOptions?: Options
-): InitializedModule;
-
+// === Overload with mandatory options ===
 function buildModule<
   InitializedModule extends Module,
   Options extends ModuleOptions,
-  InitializedExtension extends Extension<InitializedModule>
+  InitializedExtension extends Extension<InitializedModule> = object,
+  ExtensionOptions extends ModuleOptions = object
 >(
-  module: ModuleInitializer<InitializedModule, Options>,
-  moduleOptions: Options
-): InitializedModule & InitializedExtension;
-
-function buildModule<
-  InitializedModule extends Module,
-  Options extends ModuleOptions,
-  InitializedExtension extends Extension<InitializedModule>,
-  ExtensionOptions extends ModuleOptions
->(
-  module: ModuleInitializer<InitializedModule, Options>,
-  moduleOptions: Options,
-  extension?:
-    | ExtensionInitializer<
-        InitializedModule,
-        InitializedExtension,
-        ExtensionOptions
-      >
-    | InitializedExtension
-): InitializedModule & InitializedExtension;
-
-function buildModule<
-  InitializedModule extends Module,
-  Options extends ModuleOptions,
-  InitializedExtension extends Extension<InitializedModule>,
-  ExtensionOptions extends ModuleOptions
->(
-  module: ModuleInitializer<InitializedModule, Options>,
+  module: ModuleInitializerWithMandatoryOptions<InitializedModule, Options>,
   moduleOptions: Options,
   extension?:
     | ExtensionInitializer<
@@ -60,6 +28,26 @@ function buildModule<
   extensionOptions?: ExtensionOptions
 ): InitializedModule & InitializedExtension;
 
+// === Overload with optional options ===
+function buildModule<
+  InitializedModule extends Module,
+  Options extends ModuleOptions = object,
+  InitializedExtension extends Extension<InitializedModule> = object,
+  ExtensionOptions extends ModuleOptions = object
+>(
+  module: ModuleInitializer<InitializedModule, Options>,
+  moduleOptions?: Options,
+  extension?:
+    | ExtensionInitializer<
+        InitializedModule,
+        InitializedExtension,
+        ExtensionOptions
+      >
+    | InitializedExtension,
+  extensionOptions?: ExtensionOptions
+): InitializedModule & InitializedExtension;
+
+// === Implementation ===
 /**
  * Build module with extension.
  * Provide a module factory function and an extension object.
@@ -67,12 +55,12 @@ function buildModule<
  */
 function buildModule<
   InitializedModule extends Module,
-  Options extends ModuleOptions,
-  InitializedExtension extends Extension<InitializedModule>,
-  ExtensionOptions extends ModuleOptions
+  Options extends ModuleOptions = object,
+  InitializedExtension extends Extension<InitializedModule> = object,
+  ExtensionOptions extends ModuleOptions = object
 >(
   module: ModuleInitializer<InitializedModule, Options>,
-  moduleOptions: Options,
+  moduleOptions?: Options,
   extension?:
     | ExtensionInitializer<
         InitializedModule,